requirements.txt
streamlit
pandas
pandas-ta
yfinance
alpaca-trade-api
backtesting
matplotlib
pyarrow
python-dotenv